mcp_client/py_src/pyproject.toml
[project]
name = "py-src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "flask-cors>=6.0.1",
    "flask[async]>=3.1.0",
    "gunicorn>=23.0.0",
    "openai-agents>=0.0.12",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
